// Online_Pharmacy_and_Delivery_Frontend/src/shared/types/api/admin.d.ts
export interface components {
    schemas: {
// ...
        MedicineRequestDto: {
            name: string;
            genericName?: string;
            manufacturer?: string;
            dosageForm?: string;
            strength?: string;
            /** Format: double */
            price: number;
            /** Format: double */
            mrp?: number;
            /** Format: int32 */
            stock: number;
            requiresPrescription?: boolean;
            /** Format: int64 */
            categoryId: number;
            imageUrl?: string;
            description?: string;
            expiryDate: string;
            sku?: string;
        };
        ApiResponseMedicineResponseDto: {
            success?: boolean;
            message?: string;
            data?: components["schemas"]["MedicineResponseDto"];
            timestamp?: string;
        };
        MedicineResponseDto: {
            /** Format: int64 */
            id?: number;
            name?: string;
            genericName?: string;
            manufacturer?: string;
            dosageForm?: string;
            strength?: string;
            /** Format: double */
            price?: number;
            /** Format: double */
            mrp?: number;
            /** Format: int32 */
            stock?: number;
            requiresPrescription?: boolean;
            expiryDate?: string;
            /** Format: int64 */
            categoryId?: number;
            categoryName?: string;
            imageUrl?: string;
            description?: string;
            sku?: string;
            createdAt?: string;
            updatedAt?: string;
            active?: boolean;
        };
        CategoryRequestDto: {
            name: string;
            description?: string;
            imageUrl?: string;
        };
        ApiResponseCategoryResponseDto: {
            success?: boolean;
            message?: string;
            data?: components["schemas"]["CategoryResponseDto"];
            timestamp?: string;
        };
        CategoryResponseDto: {
            /** Format: int64 */
            id?: number;
            name?: string;
            description?: string;
            imageUrl?: string;
            createdAt?: string;
            active?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}